requirements.txt
Django==2.1.7
django-rest-auth==0.9.3
djangorestframework==3.9.1
pytz==2018.9
six==1.12.0
